+++ src/views/TaskMapGraphView.tsx
@@ -5,7 +5,6 @@ import ReactFlow, {
   useEdgesState,
   addEdge,
   useReactFlow,
-  ReactFlowProvider,
 } from "reactflow";
 import { Notice } from "obsidian";
 import { useApp } from "src/hooks/hooks";
@@ -33,6 +32,10 @@ const ALL_STATUSES: TaskStatus[] = ["todo", "in_progress", "done", "canceled"];
 
 interface TaskMapGraphViewProps {
   settings: TasksMapSettings;
+  selectedTags: string[];
+  setSelectedTags: React.Dispatch<React.SetStateAction<string[]>>;
+  selectedStatuses: TaskStatus[];
+  setSelectedStatuses: React.Dispatch<React.SetStateAction<TaskStatus[]>>;
 }
 
 // Helper function to filter tasks
@@ -63,39 +66,78 @@ const getFilteredNodeIds = (
   return filtered.map((task) => task.id);
 };
 
-// Inner component that uses ReactFlow hooks (must be inside ReactFlowProvider)
-interface TaskMapGraphInnerProps {
-  settings: TasksMapSettings;
-  tasks: Task[];
-  selectedTags: string[];
-  selectedStatuses: TaskStatus[];
-  setSelectedTags: React.Dispatch<React.SetStateAction<string[]>>;
-  setSelectedStatuses: React.Dispatch<React.SetStateAction<TaskStatus[]>>;
-  reloadTasks: () => void;
-  allTags: string[];
-  updateTaskTags: (taskId: string, newTags: string[]) => void;
-}
-
-function TaskMapGraphInner({
+export default function TaskMapGraphView({
   settings,
-  tasks,
   selectedTags,
-  selectedStatuses,
   setSelectedTags,
+  selectedStatuses,
   setSelectedStatuses,
-  reloadTasks,
-  allTags,
-  updateTaskTags,
-}: TaskMapGraphInnerProps) {
+}: TaskMapGraphViewProps) {
   const app = useApp();
   const vault = app.vault;
   const [nodes, setNodes, onNodesChange] = useNodesState([]);
   const [edges, setEdges, onEdgesChange] = useEdgesState([]);
+  const [tasks, setTasks] = React.useState<Task[]>([]);
   const [selectedEdge, setSelectedEdge] = React.useState<string | null>(null);
   const reactFlowInstance = useReactFlow();
 
+  // Maintain a live registry of tags per task for efficient allTags computation
+  const [taskTagsRegistry, setTaskTagsRegistry] = React.useState<
+    Map<string, string[]>
+  >(new Map());
+
+  const allTags = useMemo(() => {
+    const tagFrequency = new Map<string, number>();
+    taskTagsRegistry.forEach((tags) => {
+      tags.forEach((tag) => {
+        tagFrequency.set(tag, (tagFrequency.get(tag) || 0) + 1);
+      });
+    });
+    return Array.from(tagFrequency.keys()).sort((a, b) => {
+      const freqDiff = (tagFrequency.get(b) || 0) - (tagFrequency.get(a) || 0);
+      if (freqDiff !== 0) return freqDiff;
+      return a.localeCompare(b, undefined, { sensitivity: "base" });
+    });
+  }, [taskTagsRegistry]);
+
+  const reloadTasks = useCallback(() => {
+    const newTasks = getAllTasks(app);
+    setTasks(newTasks);
+    const newRegistry = new Map<string, string[]>();
+    newTasks.forEach((task) => {
+      newRegistry.set(task.id, task.tags);
+    });
+    setTaskTagsRegistry(newRegistry);
+    new Notice("Tasks reloaded");
+  }, [app]);
+
+  const updateTaskTags = useCallback((taskId: string, newTags: string[]) => {
+    setTaskTagsRegistry((prevRegistry) => {
+      const newRegistry = new Map(prevRegistry);
+      newRegistry.set(taskId, newTags);
+      return newRegistry;
+    });
+  }, []);
+
   useEffect(() => {
-    new Notice("Updating graph...", 1000);
+    // Wait for a short moment to ensure vault is ready
+    const timeoutId = window.setTimeout(() => {
+      reloadTasks();
+    }, 1000);
+
+    return () => window.clearTimeout(timeoutId);
+  }, [reloadTasks]);
+
+  // Update tag registry when tasks change
+  useEffect(() => {
+    const newRegistry = new Map<string, string[]>();
+    tasks.forEach((task) => {
+      newRegistry.set(task.id, task.tags);
+    });
+    setTaskTagsRegistry(newRegistry);
+  }, [tasks]);
+
+  useEffect(() => {
     let newNodes = createNodesFromTasks(
       tasks,
       settings.layoutDirection,
@@ -273,93 +315,5 @@ function TaskMapGraphInner({
         {selectedEdge && <DeleteEdgeButton onDelete={onDeleteSelectedEdge} />}
       </div>
     </TagsContext.Provider>
-  );
-}
-
-// Outer component that manages filter state and keys the ReactFlowProvider
-export default function TaskMapGraphView({ settings }: TaskMapGraphViewProps) {
-  const app = useApp();
-  const [tasks, setTasks] = React.useState<Task[]>([]);
-  const [selectedTags, setSelectedTags] = React.useState<string[]>([]);
-  const [selectedStatuses, setSelectedStatuses] = React.useState<TaskStatus[]>([
-    ...ALL_STATUSES,
-  ]);
-
-  // Maintain a live registry of tags per task for efficient allTags computation
-  const [taskTagsRegistry, setTaskTagsRegistry] = React.useState<
-    Map<string, string[]>
-  >(new Map());
-
-  const allTags = useMemo(() => {
-    const tagFrequency = new Map<string, number>();
-    taskTagsRegistry.forEach((tags) => {
-      tags.forEach((tag) => {
-        tagFrequency.set(tag, (tagFrequency.get(tag) || 0) + 1);
-      });
-    });
-    return Array.from(tagFrequency.keys()).sort((a, b) => {
-      const freqDiff = (tagFrequency.get(b) || 0) - (tagFrequency.get(a) || 0);
-      if (freqDiff !== 0) return freqDiff;
-      return a.localeCompare(b, undefined, { sensitivity: "base" });
-    });
-  }, [taskTagsRegistry]);
-
-  const reloadTasks = useCallback(() => {
-    const newTasks = getAllTasks(app);
-    setTasks(newTasks);
-    const newRegistry = new Map<string, string[]>();
-    newTasks.forEach((task) => {
-      newRegistry.set(task.id, task.tags);
-    });
-    setTaskTagsRegistry(newRegistry);
-    new Notice("Tasks reloaded");
-  }, [app]);
-
-  const updateTaskTags = useCallback((taskId: string, newTags: string[]) => {
-    setTaskTagsRegistry((prevRegistry) => {
-      const newRegistry = new Map(prevRegistry);
-      newRegistry.set(taskId, newTags);
-      return newRegistry;
-    });
-  }, []);
-
-  useEffect(() => {
-    // Wait for a short moment to ensure vault is ready
-    const timeoutId = window.setTimeout(() => {
-      reloadTasks();
-    }, 1000);
-
-    return () => window.clearTimeout(timeoutId);
-  }, [reloadTasks]);
-
-  // Update tag registry when tasks change
-  useEffect(() => {
-    const newRegistry = new Map<string, string[]>();
-    tasks.forEach((task) => {
-      newRegistry.set(task.id, task.tags);
-    });
-    setTaskTagsRegistry(newRegistry);
-  }, [tasks]);
-
-  // Key the ReactFlowProvider on filter state to force complete remount
-  const providerKey = useMemo(
-    () => `${selectedTags.join(",")}-${selectedStatuses.join(",")}`,
-    [selectedTags, selectedStatuses]
-  );
-
-  return (
-    <ReactFlowProvider key={providerKey}>
-      <TaskMapGraphInner
-        settings={settings}
-        tasks={tasks}
-        selectedTags={selectedTags}
-        selectedStatuses={selectedStatuses}
-        setSelectedTags={setSelectedTags}
-        setSelectedStatuses={setSelectedStatuses}
-        reloadTasks={reloadTasks}
-        allTags={allTags}
-        updateTaskTags={updateTaskTags}
-      />
-    </ReactFlowProvider>
   );
 }
